refactor: files related helper functions (#688)
